Added Zombie AI/Working Radio
Zombies chase AJ (without animation) and on overlap causes the level to restart (AJ died). Brian causes zombies to be destroyed on overlap (maybe add attack button later). Approaching the radio causes a button prompt to show for left click and doing so turns on the music and stops the zombies from moving while it's on.